==> LLM/Data/1.0_urls_google.ipynb <==
import os
from googleapiclient.discovery import build
from dotenv import load_dotenv
import pickle
import random
import collections
from tqdm import tqdm

random.seed(333)

load_dotenv()

api_key = os.environ['GOOGLE_API']
cse_id = os.environ['GOOGLE_CSE']

# create a search query
def google_search(search_term, api_key, cse_id, **kwargs):
    
    """
    Creates a search link for the custom Google search.
    """
    
    service = build("customsearch", "v1", developerKey=api_key)
    res = service.cse().list(q=search_term, cx=cse_id, **kwargs).execute()
    return res['items']

# search_results = google_search(palm_species[0], api_key=api_key, cse_id=cse_id)

queries = [
    "",
    "description",
    "diagnosis",
    "attributes"
]

## PALMS

folder = "Species/"

palm_species = pickle.load(open(F"{folder}palm_species.pkl", 'rb'))
# palm_species_random = random.sample(palm_species, 50)


try:
    folder = "Urls/"
    palm_species_google = pickle.load(open(F"{folder}google_urls_palms.pkl", 'rb'))
    palm_species_urls = pickle.load(open(F"{folder}urls_palms.pkl", 'rb'))
    palms_done = list(palm_species_urls.keys())
except:
    palm_species_urls = collections.defaultdict(list)
    palm_species_google = collections.defaultdict(list)
    palms_done = []

palms_done

for palm in tqdm(palm_species):

    if palm in palms_done:
        print(palm, "done")
        continue

    for query in queries:
        search_query = F"{palm} {query}"

        # print(search_query)

        # Search results (10 per search)
        search_results = google_search(search_query, api_key=api_key, cse_id=cse_id)
        # Google Urls
        palm_species_google[palm] = search_results
        # Just the links
        for result in search_results:
            palm_species_urls[palm].append(result['link'])


folder = "Urls/"

with open(F"{folder}urls_palms.pkl", 'wb') as f:
    pickle.dump(palm_species_urls, f)

with open(F"{folder}google_urls_palms.pkl", 'wb') as f:
    pickle.dump(palm_species_google, f)


folder = "Urls/"

with open(F"{folder}urls_palms.pkl", 'wb') as f:
    pickle.dump(palm_species_urls, f)

with open(F"{folder}google_urls_palms.pkl", 'wb') as f:
    pickle.dump(palm_species_google, f)


## CARIBBEAN

folder = "Species/"

caribbean_species = pickle.load(open(F"{folder}caribbean_species.pkl", 'rb'))

try:
    folder = "Urls/"
    caribbean_species_google = pickle.load(open(F"{folder}google_urls_caribbean.pkl", 'rb'))
    caribbean_species_urls = pickle.load(open(F"{folder}urls_caribbean.pkl", 'rb'))
    caribbean_done = list(caribbean_species_urls.keys())
except:
    caribbean_species_urls = collections.defaultdict(list)
    caribbean_species_google = collections.defaultdict(list)
    caribbean_done = []

caribbean_done

for caribbean in tqdm(caribbean_species):

    if caribbean in caribbean_done:
        print(caribbean, "done")
        continue

    for query in queries:
        search_query = F"{caribbean} {query}"

        # print(search_query)

        # Search results (10 per search)
        search_results = google_search(search_query, api_key=api_key, cse_id=cse_id)
        # Google Urls
        caribbean_species_google[caribbean] = search_results
        # Just the links
        for result in search_results:
            caribbean_species_urls[caribbean].append(result['link'])


folder = "Urls/"

with open(F"{folder}urls_caribbean.pkl", 'wb') as f:
    pickle.dump(caribbean_species_urls, f)

with open(F"{folder}google_urls_caribbean.pkl", 'wb') as f:
    pickle.dump(caribbean_species_google, f)

folder = "../../../data/OpenAI/Urls/"

with open(F"{folder}urls_caribbean.pkl", 'wb') as f:
    pickle.dump(caribbean_species_urls, f)

with open(F"{folder}google_urls_caribbean.pkl", 'wb') as f:
    pickle.dump(caribbean_species_google, f)

## West Africa

folder = "Species/"

west_species = pickle.load(open(F"{folder}west_species.pkl", 'rb'))

try:
    folder = "Urls/"
    west_species_google = pickle.load(open(F"{folder}google_urls_west.pkl", 'rb'))
    west_species_urls = pickle.load(open(F"{folder}urls_west.pkl", 'rb'))
    west_done = list(west_species_urls.keys())
except:
    west_species_urls = collections.defaultdict(list)
    west_species_google = collections.defaultdict(list)
    west_done = []

west_done

for west in tqdm(west_species):

    if west in westt_done:
        print(west, "done")
        continue

    for query in queries:
        search_query = F"{west} {query}"

        # print(search_query)

        # Search results (10 per search)
        search_results = google_search(search_query, api_key=api_key, cse_id=cse_id)
        # Google Urls
        west_species_google[west] = search_results
        # Just the links
        for result in search_results:
            west_species_urls[west].append(result['link'])


folder = "Urls/"

with open(F"{folder}urls_west.pkl", 'wb') as f:
    pickle.dump(west_species_urls, f)

with open(F"{folder}google_urls_west.pkl", 'wb') as f:
    pickle.dump(west_species_google, f)

folder = "Urls/"

with open(F"{folder}urls_west.pkl", 'wb') as f:
    pickle.dump(west_species_urls, f)

with open(F"{folder}google_urls_west.pkl", 'wb') as f:
    pickle.dump(west_species_google, f)
