==> LLM/Data/3.1_dataframe_to_traits.ipynb <==
import pandas as pd
import json


### Dataframes

root = "DataFrames/"

file = "DF_Caribbean.csv"
df_Caribbean = pd.read_csv(F"{root}{file}", header=[0, 1], index_col=0)\
    .rename_axis('Species', axis='index')

file = "DF_Palm.csv"
df_Palm = pd.read_csv(F"{root}{file}", header=[0, 1], index_col=0)\
    .rename_axis('Species', axis='index')

file = "DF_West.csv"
df_West = pd.read_csv(F"{root}{file}", header=[0, 1], index_col=0)\
    .rename_axis('Species', axis='index')


def create_trait_dict(multi_index):
    # Create an empty dictionary
    result_dict = {}

    # Iterate through the MultiIndex object
    for idx in multi_index:
        key = idx[0]
        value = idx[1]
        if key not in result_dict:
            result_dict[key] = []
        # Error from Andre
        if 'NA' in value:
            continue
        result_dict[key].append(value)
    
    return result_dict

root = "Traits/"

for (name, df) in zip(
            ["Caribbean", "Palm", "West"], 
            [df_Caribbean, df_Palm, df_West]
            ):
    res = create_trait_dict(df.columns)

    with open(F"{root}{name}.json", 'w') as fp:
        json.dump(res, fp)
